requirements.txt
discord.py~=1.4.0a
python-dotenv~=0.13.0
wikipedia~=1.4.0
emoji~=0.5.4
Flask~=1.1.2